source/requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.29
defusedxml==0.4.1
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.30.0
django-crispy-forms==1.6.1
django-guardian==1.4.6
django-storages==1.5.2
docutils==0.13.1
geoip2==2.5.0
gunicorn==19.7.1
jmespath==0.9.2
maxminddb==1.3.0
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.7.0
s3transfer==0.1.10
six==1.10.0
